fix(examples): update imports for relocated demo packages
The demos moved under examples/demos/ but five files still imported
from the old examples.<pkg> module paths, breaking them when run.
